SurfaceFlinger: keep auto-refresh in sync between producer and consumer
BufferQueueLayer updates mAutoRefresh when calling acquireBuffer(),
which means that the latest value set by the producer
(android::Surface) may not always be correctly propagated to the
consumer. This can break things that expect them to be in sync.

The auto-refresh flag can get out of sync in the following two cases
(both require shared buffer mode to be set as well):
1. Producer disables auto refresh. At this point acquireBuffer() will
return NO_BUFFER_AVAILABLE, so there will be no BufferItem to read
the value from.
2. Producer enables auto refresh, but doesn't call queueBuffer()
afterwards. In this case hasReadyFrame() continues to return false,
so SurfaceFlinger doesn't call latchBuffer() which is needed to
update mAutoRefresh.

This diff adds a new method to IGraphicBufferConsumer that returns the
value of the auto-refresh flag directly. Furthermore, when the producer
sets the flag, it notifies SurfaceFlinger (via IConsumerListener)
to refresh the layer and fetch the latest value.

Test: BufferQueue test added.
Change-Id: Ic336f7cf03d04738d7247d266302b9f5e1fabdb0
Signed-off-by: fadlyas07 <mhmmdfdlyas@gmail.com>